Move debounce to utils and clear search on named view
