+++ src/js/components/FormField/FormField.js
@@ -4,7 +4,6 @@ import React, {
   forwardRef,
   useContext,
   useState,
-  useEffect,
 } from 'react';
 import styled, { ThemeContext } from 'styled-components';
 import { defaultProps } from '../../default-props';
@@ -14,6 +13,7 @@ import {
   shouldKeepFocus,
   withinDropPortal,
   PortalContext,
+  useDebounce,
 } from '../../utils';
 import { focusStyle } from '../../utils/styles';
 import { parseMetricToNum } from '../../utils/mixins';
@@ -154,19 +154,6 @@ const Input = ({ component, disabled, invalid, name, onChange, ...rest }) => {
       {...extraProps}
     />
   );
-};
-
-const useDebounce = () => {
-  const [func, setFunc] = useState();
-  const theme = useContext(ThemeContext) || defaultProps.theme;
-
-  useEffect(() => {
-    let timer;
-    if (func) timer = setTimeout(() => func(), theme.global.debounceDelay);
-    return () => clearTimeout(timer);
-  }, [func, theme.global.debounceDelay]);
-
-  return setFunc;
 };
 
 const FormField = forwardRef(
