+++ src/js/utils/use-debounce.js
@@ -1,0 +1,17 @@
+import { useContext, useEffect, useState } from 'react';
+import { ThemeContext } from 'styled-components';
+
+export const useDebounce = (debounceDelay) => {
+  const [func, setFunc] = useState();
+  const theme = useContext(ThemeContext) || defaultProps.theme;
+  const delay = debounceDelay || theme.global.debounceDelay;
+  useEffect(() => {
+    let timer;
+    if (func) timer = setTimeout(() => func(), delay);
+    return () => clearTimeout(timer);
+  }, [func, delay]);
+
+  return setFunc;
+};
+
+export default useDebounce;
